+++ docs/CHANGELOG.rst
@@ -3,10 +3,10 @@ Appendix C. Changelog
 =====================
 :Info: This is the changelog for Cheqlist.
 :Author: Chris Warrick <chris@chriswarrick.com>
-:Copyright: © 2015, Chris Warrick.
+:Copyright: © 2015-2016, Chris Warrick.
 :License: BSD (see /LICENSE or :doc:`Appendix B <LICENSE>`.)
-:Date: 2015-10-25
-:Version: 0.1.6
+:Date: 2016-07-15
+:Version: 0.2.0
 
 .. index:: CHANGELOG
 
@@ -18,6 +18,14 @@ More information can be found on GitHub in the `releases section
 
 Version History
 ===============
+
+0.2.0
+    * Display exceptions in message boxes
+    * More formatting: underline and strikeout
+    * Remove sample items
+    * Split into two toolbars (Main and Edit)
+    * Make it possible to ignore struck-out tasks
+    * Add right-click menus
 
 0.1.6
     * Fix updating last used directory